+++ packages/class-core/src/serializer.ts
@@ -21,6 +21,8 @@ import {
     Containment,
     Enumeration,
     Feature,
+    isRef,
+    isReferenceToSet,
     isUnresolvedReference,
     LionWebVersions,
     Node,
@@ -120,8 +122,8 @@ export const propertyValueSerializerWith = (configuration?: PropertyValueSeriali
     return {
         serializeValue: (value: unknown, property: Property) => {
             const { type } = property
-            if (isUnresolvedReference(type)) {
-                reportIssue(`can't serialize value of property "${property.name}" (on classifier "${property.classifier.name}" in language "${property.classifier.language.name}") having unresolved type: ${value}`)
+            if (!isRef(type)) {
+                reportIssue(`can't serialize value of property "${property.name}" (on classifier "${property.classifier.name}" in language "${property.classifier.language.name}") having ${isReferenceToSet(type) ? `an unset type` : `unresolved type: ${value}${isUnresolvedReference(type) && type.resolveInfo !== undefined ? ` -> ${type.resolveInfo}` : ``}`}`)
                 return null
             }
             if (type instanceof PrimitiveType) {