+++ packages/delta-protocol-common/src/translators/delta-to-event.ts
@@ -37,7 +37,6 @@ import {
     IDelta,
     idFrom,
     INodeBase,
-    nodeBaseReader,
     NoOpDelta,
     PartitionAddedDelta,
     PartitionDeletedDelta,
@@ -82,6 +81,7 @@ import {
     ReferenceChangedEvent,
     ReferenceDeletedEvent
 } from "../payload/index.js"
+import { resolveInfoFrom } from "./ref-util.js"
 
 
 const allIdsOfDescendantsFrom = (node: INodeBase) =>
@@ -354,7 +354,7 @@ export const deltaToEventTranslator = (
                     reference: metaPointerForFeature(delta.reference),
                     index: delta.index,
                     newReference: idFrom(delta.newReference),
-                    newResolveInfo: nodeBaseReader.resolveInfoFor!(delta.newReference!, delta.reference)!
+                    newResolveInfo: resolveInfoFrom(delta.newReference, delta.reference)
                 })
             }
             if (delta instanceof ReferenceDeletedDelta) {
@@ -363,7 +363,7 @@ export const deltaToEventTranslator = (
                     reference: metaPointerForFeature(delta.reference),
                     index: delta.index,
                     deletedReference: idFrom(delta.deletedReference),
-                    deletedResolveInfo: nodeBaseReader.resolveInfoFor!(delta.deletedReference!, delta.reference)!
+                    deletedResolveInfo: resolveInfoFrom(delta.deletedReference, delta.reference)
                 })
             }
             if (delta instanceof ReferenceChangedDelta) {
@@ -372,9 +372,9 @@ export const deltaToEventTranslator = (
                     reference: metaPointerForFeature(delta.reference),
                     index: delta.index,
                     oldReference: idFrom(delta.oldReference),
-                    oldResolveInfo: nodeBaseReader.resolveInfoFor!(delta.oldReference!, delta.reference)!,
+                    oldResolveInfo: resolveInfoFrom(delta.oldReference, delta.reference),
                     newReference: idFrom(delta.newReference),
-                    newResolveInfo: nodeBaseReader.resolveInfoFor!(delta.newReference!, delta.reference)!
+                    newResolveInfo: resolveInfoFrom(delta.newReference, delta.reference)
                 })
             }
             if (delta instanceof CompositeDelta) {